Add support for deals using a data segment index
A deal may, rather than a single car, be a series of concatinated car files with a segment index at the end containing an inclusion proof.
When such an index is found, we will allow reading multiple cars and taking valid cids found in all of them as the contents of the overall deal.